refactor fips code
